movendo coisas do details pra category, evitando conflito
